requirements.txt
numpy
pandas
pillow
Flask
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
requests
bs4
python-dotenv
pyautogui
selenium
twilio
